feat [Pautista/PautaService]: adicionando atributo id_sapiens a tabela de pautistas e ajuste no metodo processar Pautas para pegar este id do banco

+++ src/main/java/com/agu/gestaoescalabackend/client/request/TarefaLoteRequest.java
@@ -23,7 +23,7 @@ public class TarefaLoteRequest {
     @NotNull
     private int setorResponsavel; //vem do front
     @NotNull
-    private int usuarioResponsavel; //back
+    private Long usuarioResponsavel; //back
 
     @NotNull
     private List<ProcessoJudicialDTO> listaProcessosJudiciais;
